Update README.md
I thought up of this project during lockdown when all the gyms were closed and I couldn't exercise like I used to and wondered if there was a way to exercise correctly - with an instructor an all - while being at home. 
Not everyone has enough knowledge about workouts or even equipments and even if they do identify a few, they aren't really sure what to do with them. This algo solves it all. It can identify various equipments through your camera and thus redirect you to various videos on YouTube on how to properly use them. Along with this, when you do actually work out, it scans your body posture in live time.
It uses an algorithm called YOLOV4 which stands for ‘You only look once’ and uses CNN to scan the
equipment and identify what it is and thus redirect you to respective webpages showing how they
can be used. 

Another algorithm called POSENET will be used to view your live time position by
detecting your major joints of the body and connecting them to draw a stick figure which can then
be compared to an ideal posture and thus suggest corrections based on that